Fix infinite recursion bug
Attempting to assign the imported module to `navigator.clipboard` and then
using the `writeText` or `readText` methods would recurse infinitely, resulting
in a stack overflow.

The reason for this was the faulty attempt to read from `navigator.clipboard` if
it's available, regardless of what that binding actually references.

This commit caches `navigator.clipboard` and avoids the use of it in case it
contains a reference to the polyfill itself. Otherwise, the cached binding is
used to avoid recursion.